third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/edit/set-attribute-expected.txt

Tests that elements panel updates dom tree structure upon setting attribute.


Running: testDumpInitial
========= Original ========
  <div id="node"></div>

Running: testAttributeUpdated
===== On attribute set =====
  <div id="node" name="value"></div>

Running: testAttributeSameValueNotUpdated
===== On attribute modified (should be 'newValue') =====
  <div id="node" name="newValue"></div>

Running: testAttributeRemoved
=== On attribute removed ===
  <div id="node"></div>

Running: testSetAttributeValue
=== Set attribute value ===
  <div id="node" foo="bar"></div>

Running: testSetAttributeText
=== Set attribute as text ===
  <div id="node" foo2="baz2" foo3="baz3"></div>

Running: testRemoveAttributeAsText
=== Remove attribute as text ===
  <div id="node" foo2="baz2"></div>

Running: testSetMalformedAttributeText
Error: Could not parse value as attributes
=== Set malformed attribute as text ===
  <div id="node" foo2="baz2"></div>
